+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- fakeCUDA: add device-mars-x203.h;
 - targets/nvidia.mk: add .nv_fatbin and .nvFatBinSegment targets;
 - fakeCUDA: add fatbin.hpp, fatbin.cpp, fatbin-hip.hpp;
 - fakeCUDA: add runtime.hpp;